+++ node/cryp.h
@@ -1,12 +1,14 @@
 #ifndef CRYP_H
 #define CRYP_H
 
+#include <stddef.h>
+
 #define HASH_LEN 32
 
 struct keys;
 struct public_key;
 
-void cryp_hash(unsigned char *out, void *in, int len);
+void cryp_hash(unsigned char *out, void *in, size_t len);
 void cryp_hash_cmp(int *ret, unsigned char *h1, unsigned char *h2);
 void cryp_hash_dist(unsigned char *dist, unsigned char *h1, unsigned char *h2);
